Move sorting implementations for Native to kotlin.collections
